* Fixed patch config parsing for tests
