+++ tests/TEST_INPUTS/patch-configs/FeaturePerfCSCollection/test-patch-configuration.xml
@@ -1,5 +1,5 @@
 <config>
-	<repository>github.com/se-sic/FeaturePerfCSCollection.git</repository>
+	<repository>git@github.com:se-sic/FeaturePerfCSCollection.git</repository>
 	<patches>
 		<patch>
 			<shortname>unrestricted-range</shortname>
@@ -26,8 +26,8 @@
 			<path>bug.patch</path>
 			<include_revisions>
 				<revision_range>
-					<start_revision>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start_revision>
-					<end_revision>8ca5cc28e6746eef7340064b5d843631841bf31e</end_revision>
+					<start>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start>
+					<end>8ca5cc28e6746eef7340064b5d843631841bf31e</end>
 				</revision_range>
 			</include_revisions>
 		</patch>
@@ -39,8 +39,8 @@
 			<path>bug.patch</path>
 			<include_revisions>
 				<revision_range>
-					<start_revision>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start_revision>
-					<end_revision>8ca5cc28e6746eef7340064b5d843631841bf31e</end_revision>
+					<start>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start>
+					<end>8ca5cc28e6746eef7340064b5d843631841bf31e</end>
 				</revision_range>
 				<single_revision>
 					27f17080376e409860405c40744887d81d6b3f34
@@ -65,8 +65,8 @@
 			<path>bug.patch</path>
 			<exclude_revisions>
 				<revision_range>
-					<start_revision>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start_revision>
-					<end_revision>8ca5cc28e6746eef7340064b5d843631841bf31e</end_revision>
+					<start>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start>
+					<end>8ca5cc28e6746eef7340064b5d843631841bf31e</end>
 				</revision_range>
 			</exclude_revisions>
 		</patch>
@@ -78,8 +78,8 @@
 			<path>bug.patch</path>
 			<exclude_revisions>
 				<revision_range>
-					<start_revision>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start_revision>
-					<end_revision>8ca5cc28e6746eef7340064b5d843631841bf31e</end_revision>
+					<start>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start>
+					<end>8ca5cc28e6746eef7340064b5d843631841bf31e</end>
 				</revision_range>
 				<single_revision>
 					27f17080376e409860405c40744887d81d6b3f34
@@ -94,8 +94,8 @@
 			<path>bug.patch</path>
 			<include_revisions>
 				<revision_range>
-					<start_revision>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start_revision>
-					<end_revision>8ca5cc28e6746eef7340064b5d843631841bf31e</end_revision>
+					<start>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start>
+					<end>8ca5cc28e6746eef7340064b5d843631841bf31e</end>
 				</revision_range>
 			</include_revisions>
 			<exclude_revisions>
@@ -110,14 +110,14 @@
 			<path>bug.patch</path>
 			<include_revisions>
 				<revision_range>
-					<start_revision>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start_revision>
-					<end_revision>4300ea495e7f013f68e785fdde5c4ead81297999</end_revision>
+					<start>01f9f1f07bef22d4248e8349aba4f0c1f204607e</start>
+					<end>4300ea495e7f013f68e785fdde5c4ead81297999</end>
 				</revision_range>
 			</include_revisions>
 			<exclude_revisions>
 				<revision_range>
-					<start_revision>c051e44a973ee31b3baa571407694467a513ba68</start_revision>
-					<end_revision>a94fb35ca49719028a1c50bdbc2fb82122043f46</end_revision>
+					<start>c051e44a973ee31b3baa571407694467a513ba68</start>
+					<end>a94fb35ca49719028a1c50bdbc2fb82122043f46</end>
 				</revision_range>
 			</exclude_revisions>
 		</patch>
