+++ tests/provider/test_patch_provider.py
@@ -1,8 +1,15 @@
 import unittest
 from pathlib import Path
 
+import benchbuild as bb
+from benchbuild.source.base import target_prefix
+from benchbuild.utils.revision_ranges import _get_git_for_path
+
 from tests.helper_utils import TEST_INPUTS_DIR
-from varats.provider.patch.patch import ProjectPatches
+from varats.provider.patch.patch_provider import ProjectPatchesConfiguration
+
+from varats.project.project_util import get_local_project_git_path
+from varats.utils.git_util import ShortCommitHash
 
 
 class TestPatchProvider(unittest.TestCase):
@@ -13,10 +20,26 @@ class TestPatchProvider(unittest.TestCase):
 class TestPatchConfiguration(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
-        cls.patch_config = ProjectPatches.from_xml( Path(TEST_INPUTS_DIR/'patch-configs/test-patch-configuration.xml') )
+        patch_config = ProjectPatchesConfiguration.from_xml(Path(TEST_INPUTS_DIR / 'patch-configs/FeaturePerfCSCollection/test-patch-configuration.xml'))
+        cls.patch_config = patch_config
+
+        project_git_source = bb.source.Git(
+            remote="git@github.com:se-sic/FeaturePerfCSCollection.git",
+            local="FeaturePerfCSCollection",
+            refspec="origin/HEAD",
+            shallow=False,
+        )
+
+        project_git_source.fetch()
+
+        repo_git = _get_git_for_path(target_prefix() + "/FeaturePerfCSCollection")
+
+        cls.all_revisions = { ShortCommitHash(h) for h in repo_git('log', '--pretty=%H', '--first-parent').strip().split() }
 
     def test_unrestricted_range(self):
-        pass
+        patch = self.patch_config.get_by_shortname('unrestricted-range')
+
+        self.assertEqual(patch.valid_revisions, set(self.all_revisions))
 
     def test_included_single_revision(self):
         pass
